requirements.txt
rasterio~=1.3.10
shapely~=2.0.4
Flask~=3.0.3
pytest~=8.2.2
python-dotenv~=1.0.1
